pyproject.toml
[project]
name = "pydanticai-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "google-generativeai>=0.8.5",
    "pydantic-ai>=0.2.0",
    "python-dotenv>=1.1.0",
]
